fix: separate chain pkg from utils
